Fix some compilation problems on MacOS

+++ core/acpc-run.cpp
@@ -124,7 +124,7 @@ int main(int argc, char**argv) {
   EXPERIMENT_TAG("delta", delta);
   EXPERIMENT_TAG("rate", rate);
   EXPERIMENT_TAG("p_low", p_low);
-  EXPERIMENT_TAG("seed", seed);
+  EXPERIMENT_TAG("seed", (size_t) seed);
   EXPERIMENT_TAG("k", k);
   EXPERIMENT_TAG("batch size", h);
   EXPERIMENT_TAG("git-revision", std::string(g_GIT_SHA1));
