+++ core/gmm.cpp
@@ -164,7 +164,7 @@ int main(int argc, char **argv) {
   EXPERIMENT_TAG("p_low", p_low);
   EXPERIMENT_TAG("epsilon", epsilon);
   EXPERIMENT_TAG("delta", delta);
-  EXPERIMENT_TAG("seed", seed);
+  EXPERIMENT_TAG("seed", (size_t) seed);
   EXPERIMENT_TAG("k", k);
   EXPERIMENT_TAG("git-revision", std::string(g_GIT_SHA1));
   EXPERIMENT_TAG("theory-samples-fraction", theory_samples_fraction);
